DDF-1760 Refactored into api and impl packages. Updated content to use dynamic metacards
