+++ catalog/core/catalog-core-dynamic-impl/src/test/java/ddf/catalog/data/dynamic/Serializer.java
