+++ catalog/core/catalog-core-dynamic-impl/src/test/resources/logback.xml
