--- catalog/core/catalog-core-dynamic/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version. 
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-
- -->
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xmlns:camel="http://camel.apache.org/schema/blueprint" >
-
-    <bean id="dynamicBeanFactory" class="ddf.catalog.data.dynamic.impl.MetacardFactoryImpl" />
-
-    <service id="dynBeanFactory" ref="dynamicBeanFactory" interface="ddf.catalog.data.dynamic.api.MetacardFactory">
-    </service>
-
-    <bean id="metacardReader" class="ddf.catalog.data.dynamic.registry.MetacardReader">
-        <property name="metacardFactory" ref="dynamicBeanFactory"/>
-    </bean>
-
-    <!-- Monitor the 'etc/metadata' folder for changes, passing each file
-         received to the MetacardReader for parsing and registering. -->
-    <camelContext xmlns="http://camel.apache.org/schema/blueprint">
-      <route id="MCT_MONITOR">
-          <from uri="file://etc/metadata?noop=true&amp;delay=10000"/>
-          <to uri="bean:metacardReader?method=registerMetacardDefinition" />
-      </route>
-  </camelContext>
-
-</blueprint>
